+++ macos/QMK Toolbox/Info.plist
@@ -93,8 +93,8 @@
 	</dict>
 	<key>NSHumanReadableCopyright</key>
 	<string>Copyright 2022 QMK</string>
-	<key>NSMainNibFile</key>
-	<string>MainMenu</string>
+	<key>NSMainStoryboardFile</key>
+	<string>Main</string>
 	<key>NSPrincipalClass</key>
 	<string>NSApplication</string>
 	<key>UTExportedTypeDeclarations</key>
